requirements.txt
# Core Scientific Libraries
numpy
pandas
scikit-learn
scipy

# Deep Learning Framework
torch
torchvision
torchaudio
pytorch-lightning

# Machine Learning Tools
transformers
tokenizers
huggingface-hub
sentencepiece

# Visualization and Logging
tensorboard
matplotlib
seaborn

# Progress and Performance
tqdm
psutil

# Data Processing
pyarrow
fastparquet
python-dateutil

# Configuration and File Handling
pyyaml
json5
python-dotenv

# Optional but Recommended
joblib
typing-extensions
packaging
pillow
requests
regex
